+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 - Basic support for flat formats (see https://github.com/ehrbase/openEHR_SDK/pull/67) 
 - Web-Templates (see https://github.com/ehrbase/openEHR_SDK/pull/81)
 - Allow Native Parts in Entity Queries  (see https://github.com/ehrbase/openEHR_SDK/pull/106)
+- New endpoint for retrieving all templates
 ### Fixed
 - Error when extracting name from a template where name has more than one child (see https://github.com/ehrbase/openEHR_SDK/pull/79)
 - Enums where not correctly generated for value-sets (see https://github.com/ehrbase/openEHR_SDK/pull/92)  
@@ -13,4 +14,5 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 ## 0.3.5
 
 - patch: openEHR SDK containing the contents of EHRbase serialization, response DTOs, terminology and validation
+